Added distance metric to gain calculation and more parameters to launch files

+++ rrt_nbv_exploration/include/rrt_nbv_exploration/RneVisualizer.h
@@ -53,6 +53,6 @@ private:
 	 * @param Gain of the node
 	 */
 	void addInfoTextVisualization(const geometry_msgs::Point node_position,
-			int node, int gain);
+			int node, double gain);
 };
 }
